requirements.txt
fastapi
uvicorn
chromadb
pymongo
dnspython
sentence-transformers
numpy
markitdown
google-genai
python-multipart
jinja2
aiofiles
python-dotenv
pytest
httpx
pytest-asyncio
